adk-microservice/requirements.txt
fastapi
uvicorn
pydantic
google-adk
google-genai
google-cloud-aiplatform
python-dotenv
llama-index 